Move Docker creation to RM configs
