+++ x-pack/test/security_solution_api_integration/test_suites/detections_response/rules_management/configs/ess/rules_management.trial.config.ts
@@ -5,13 +5,21 @@
  * 2.0.
  */
 
-import { FtrConfigProviderContext } from '@kbn/test';
+import path from 'path';
+import {
+  FtrConfigProviderContext,
+  defineDockerServersConfig,
+  fleetPackageRegistryDockerImage,
+} from '@kbn/test';
 import { LOGGING_CONFIG, FLEET_PLUGIN_READY_LOG_MESSAGE_REGEXP } from '../constants';
 
 export default async function ({ readConfigFile }: FtrConfigProviderContext) {
   const functionalConfig = await readConfigFile(
     require.resolve('../../../../../config/ess/config.base.trial')
   );
+
+  const packageRegistryConfig = path.join(__dirname, './package_registry_config.yml');
+  const dockerArgs: string[] = ['-v', `${packageRegistryConfig}:/package-registry/config.yml`];
 
   return {
     ...functionalConfig.getAll(),
@@ -25,5 +33,16 @@ export default async function ({ readConfigFile }: FtrConfigProviderContext) {
         wait: FLEET_PLUGIN_READY_LOG_MESSAGE_REGEXP,
       },
     },
+    dockerServers: defineDockerServersConfig({
+      registry: {
+        enabled: true,
+        image: fleetPackageRegistryDockerImage,
+        portInContainer: 8080,
+        port: 8081,
+        args: dockerArgs,
+        waitForLogLine: 'package manifests loaded',
+        waitForLogLineTimeoutMs: 60 * 4 * 1000, // 4 minutes
+      },
+    }),
   };
 }
