+++ x-pack/test/security_solution_api_integration/test_suites/detections_response/rules_management/configs/serverless/package_registry_config.yml
@@ -1,0 +1,2 @@
+package_paths:
+  - /packages/package-storage
